refactoruing and fix test
