+++ crack-testing/Cargo.toml
@@ -18,7 +18,7 @@ workspace = ".."
 anyhow = "1.0"
 crack-types = { path = "../crack-types" }
 clap = { version = "4.5", features = ["derive"] }
-dashmap = "6.1"
+dashmap = "5.5.3"
 once_cell = "1.20"
 futures = "0.3"
 rand = "0.8"
